Improve API response handling

+++ Partners/services/speaker.service.js
@@ -3,18 +3,42 @@ const SpeakerModel = require('../models/speaker.model.js');
 const getSpeakers = async () => {
   try {
     const speakers = await SpeakerModel.find({});
-    return speakers;
+    if (!speakers) {
+      return {
+        status: 400,
+
+        message: 'Failed to retrieve speakers',
+      };
+    }
+    if (speakers.length === 0) {
+      return { status: 400, message: 'No speakers available' };
+    }
+
+    return {
+      status: 200,
+      data: speakers,
+      message: 'Retrieved all speakers successfully',
+    };
   } catch (error) {
-    return error;
+    return { status: 400, message: error.message };
   }
 };
 
 const getEventSpecificSpeakers = async (eventID) => {
   try {
     const speakers = await SpeakerModel.find({ eventID: eventID });
-    return speakers;
+
+    if (speakers.length === 0) {
+      return { status: 400, message: 'No speakers available' };
+    }
+
+    return {
+      status: 200,
+      data: speakers,
+      message: 'Retrieved all event specific speakers successfully ',
+    };
   } catch (error) {
-    return error;
+    return { status: 400, message: error.message };
   }
 };
 
@@ -26,12 +50,17 @@ const addASpeaker = async ({
   eventID,
   organizerID,
 }) => {
-  const availableSpeaker = await SpeakerModel.findOne({ fullName: fullName });
-  if (availableSpeaker) {
-    if (availableSpeaker.eventID == eventID) {
-      return "'Speaker Already Exists For the Event'";
+  try {
+    const availableSpeaker = await SpeakerModel.findOne({
+      fullName: fullName,
+      eventID: eventID,
+    });
+    if (availableSpeaker) {
+      return {
+        status: 400,
+        message: 'Speaker already exists in the event',
+      };
     }
-  } else {
     const speaker = new SpeakerModel({
       fullName: fullName,
       sessionTime: sessionTime,
@@ -41,38 +70,53 @@ const addASpeaker = async ({
       organizerID: organizerID,
     });
 
-    try {
-      const response = await speaker.save();
-      return response;
-    } catch (error) {
-      return error;
-    }
+    const response = await speaker.save();
+    return {
+      status: 200,
+      data: response,
+      message: 'Added the speaker to the event successfully',
+    };
+  } catch (error) {
+    return { status: 400, message: error.message };
   }
 };
 
-const updateSpeakerDetails = async ({ id, body }) => {
+const updateSpeakerDetails = async (id, body) => {
   try {
-    const speakerToBeUpdated = await SpeakerModel.findByIdAndUpdate(id, body);
-    if (!speakerToBeUpdated) {
-      console.log('not found');
+    const isSpeakerAvailable = await SpeakerModel.findById(id);
+
+    if (!isSpeakerAvailable) {
+      return { status: 404, message: `Speaker with id ${id} not found` };
     }
-    return 'Updated the Form';
+
+    const response = await SpeakerModel.findByIdAndUpdate(id, body);
+    return {
+      status: 200,
+      data: response,
+      message: `Updated the speaker successfully`,
+    };
   } catch (error) {
-    return error;
+    return { status: 400, message: error.message };
   }
 };
 
 const deleteSpeaker = async (id) => {
   try {
     const speakerToBeDeleted = await SpeakerModel.findById(id);
+    if (!speakerToBeDeleted) {
+      return { status: 404, message: `Speaker with id ${id} not found` };
+    } else {
+      await SpeakerModel.deleteOne({
+        _id: speakerToBeDeleted._id,
+      });
 
-    await SpeakerModel.deleteOne({
-      _id: speakerToBeDeleted._id,
-    });
-
-    return 'speaker Deleted';
+      return {
+        status: 200,
+        message: 'Deleted the Speaker successfully',
+      };
+    }
   } catch (error) {
-    return error;
+    return { status: 400, message: error.message };
   }
 };
