+++ Partners/services/volunteer.service.js
@@ -3,9 +3,23 @@ const VolunteerModel = require('../models/volunteer.model.js');
 const getVolunteers = async () => {
   try {
     const volunteers = await VolunteerModel.find({});
-    return volunteers;
+    if (!volunteers) {
+      return {
+        status: 400,
+
+        message: 'Failed to retrieve Volunteers',
+      };
+    }
+    if (volunteers.length === 0) {
+      return { status: 400, message: 'No volunteers available' };
+    }
+    return {
+      status: 200,
+      data: volunteers,
+      message: 'Retrieved all volunteers successfully',
+    };
   } catch (error) {
-    return error;
+    return { status: 400, message: error.message };
   }
 };
 
@@ -18,14 +32,18 @@ const applyToAnOpportunity = async ({
   userID,
   opportunityID,
 }) => {
-  const availableVolunteer = await VolunteerModel.findOne({
-    fullName: fullName,
-  });
-  if (availableVolunteer) {
-    if (availableVolunteer.opportunityID == opportunityID) {
-      return 'Volunteer Already Exists For the Event';
+  try {
+    const availableVolunteer = await VolunteerModel.findOne({
+      userID: userID,
+      opportunityID: opportunityID,
+    });
+    if (availableVolunteer) {
+      return {
+        status: 400,
+
+        message: 'Volunteer already has applied to the opportunity',
+      };
     }
-  } else {
     const volunteer = new VolunteerModel({
       fullName: fullName,
       email: email,
@@ -36,27 +54,35 @@ const applyToAnOpportunity = async ({
       opportunityID: opportunityID,
     });
 
-    try {
-      const response = await volunteer.save();
-      return response;
-    } catch (error) {
-      return error;
-    }
+    const response = await volunteer.save();
+
+    return {
+      status: 200,
+      data: response,
+      message: 'Updated the volunteer application successfully',
+    };
+  } catch (error) {
+    return { status: 400, message: error.message };
   }
 };
 
-const updateVolunteerApplication = async ({ id, body }) => {
+const updateVolunteerApplication = async (id, body) => {
   try {
-    const applicationToBeUpdate = await VolunteerModel.findByIdAndUpdate(
-      id,
-      body
-    );
-    if (!applicationToBeUpdate) {
-      console.log('not found');
+    const isApplicationAvailable = await VolunteerModel.findById(id);
+
+    if (!isApplicationAvailable) {
+      return { status: 404, message: `Volunteer with id ${id} not found` };
     }
-    return 'Updated the Form';
+
+    const response = await VolunteerModel.findByIdAndUpdate(id, body);
+
+    return {
+      status: 200,
+      data: response,
+      message: `Updated Volunteer application successfully`,
+    };
   } catch (error) {
-    return error;
+    return { status: 400, message: error.message };
   }
 };
 
@@ -64,13 +90,20 @@ const deleteVolunteerApplication = async (id) => {
   try {
     const volunteerToBeDeleted = await VolunteerModel.findById(id);
 
-    await VolunteerModel.deleteOne({
-      _id: volunteerToBeDeleted._id,
-    });
+    if (!volunteerToBeDeleted) {
+      return { status: 404, message: `Volunteer with id ${id} not found` };
+    } else {
+      await VolunteerModel.deleteOne({
+        _id: volunteerToBeDeleted._id,
+      });
 
-    return 'Volunteer Deleted';
+      return {
+        status: 200,
+        message: 'Deleted volunteer application successfully',
+      };
+    }
   } catch (error) {
-    return error;
+    return { status: 400, message: error.message };
   }
 };
